Fixed import issue on Python3 and Added tests
Changed the import syntax such that modules can be imported properly on Python3.
Added tests on pipeline_config module and added 'mock' in requirements.txt and setup.py because it is used in the test script.
Also replaced the print statement in pipeliner and pipeline_config modules to logging, but the print statements in modules in core haven't been replace.